change ByteBuffer to BitBuffer

+++ DoubleSocket.Example.Server/Player.cs
@@ -4,12 +4,12 @@ namespace DoubleSocket.Example.Server {
 	public class Player {
 		public IDoubleServerClient ServerClient { get; }
 		public byte Id { get; }
-		public int Color { get; }
+		public byte Color { get; }
 		public byte X { get; set; } = byte.MaxValue;
 		public byte Y { get; set; } = byte.MaxValue;
-		public ushort NewestPacketTimestamp;
+		public uint NewestPacketTimestamp;
 
-		public Player(IDoubleServerClient serverClient, byte id, int color) {
+		public Player(IDoubleServerClient serverClient, byte id, byte color) {
 			ServerClient = serverClient;
 			Id = id;
 			Color = color;
